+++ services/director-v2/src/simcore_service_director_v2/utils/comp_scheduler.py
@@ -1,3 +1,5 @@
+from typing import TypeAlias
+
 from models_library.projects_state import RunningState
 from pydantic import PositiveInt
 
@@ -31,4 +33,4 @@ COMPLETED_STATES: set[RunningState] = {
 }
 
 
-Iteration = PositiveInt
+Iteration: TypeAlias = PositiveInt
